requirements.txt
https://github.com/heshameraqi/labelImg_OBB
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
distutils-pytest==0.2.1
filelock==3.15.4
fonttools==4.53.0
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
iniconfig==2.0.0
intel-openmp==2021.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
lapx==0.5.9.post1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
mysql-connector==2.2.9
mysql-connector-python==8.4.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pluggy==1.5.0
psutil==6.0.0
psycopg2==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PySimpleGUI==5.0.6
pytest==8.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
scipy==1.14.0
seaborn==0.13.2
setuptools==70.1.1
shapely==2.0.4
sip==6.8.6
six==1.16.0
smmap==5.0.1
supervision==0.21.0
sympy==1.12.1
tbb==2021.13.0
thop==0.1.1.post2209072238
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wheel==0.43.0